+++ plugins/remote-config/api/api.js
@@ -361,6 +361,11 @@ plugins.setConfigs("remote-config", {
                 }
                 return common.returnMessage(params, 500, message);
             }
+
+            plugins.dispatch("/i/remote-config/add-parameter", {
+                params: parameter
+            });
+
             if (params.internal) {
                 return true;
             }
@@ -823,6 +828,10 @@ plugins.setConfigs("remote-config", {
                 return common.returnMessage(params, 500, message);
             }
 
+            plugins.dispatch("/i/remote-config/update-parameter", {
+                params: parameter
+            });
+
             return common.returnMessage(params, 200);
         });
     }
@@ -945,6 +954,9 @@ plugins.setConfigs("remote-config", {
                 common.outDb.collection(collectionName).remove({_id: common.outDb.ObjectID(parameterId)}, function(removeErr) {
                     if (!removeErr) {
                         plugins.dispatch("/systemlogs", {params: params, action: "rc_parameter_removed", data: parameter});
+                        plugins.dispatch("/i/remote-config/remove-parameter", {
+                            params: parameter
+                        });
                         return common.returnMessage(params, 200, 'Success');
                     }
 
@@ -1023,6 +1035,9 @@ plugins.setConfigs("remote-config", {
             }
 
             var conditionId = result && result[1] || null;
+            plugins.dispatch("/i/remote-config/add-condition", {
+                params: condition
+            });
             if (params.internal) {
                 return conditionId;
             }
@@ -1193,6 +1208,10 @@ plugins.setConfigs("remote-config", {
                 return common.returnMessage(params, 500, message);
             }
 
+            plugins.dispatch("/i/remote-config/update-condition", {
+                params: condition
+            });
+
             return common.returnMessage(params, 200);
         });
 
@@ -1236,6 +1255,13 @@ plugins.setConfigs("remote-config", {
 
         async.parallel(asyncTasks, function(err) {
             if (!err) {
+                plugins.dispatch("/i/remote-config/remove-condition", {
+                    params: {
+                        conditionId: conditionId,
+                        appId: appId,
+                        ts: Date.now()
+                    }
+                });
                 return common.returnMessage(params, 200, 'Success');
             }